Fix products / compute more products
Before, we were just computing Ext^0 x Ext^0. Then we tried to compute
Ext^0 x Ext^{-1} as well. I'm not sure if this was wrong, but I realized
that we need to compute all Ext^0 x Ext^m. A new argument was added (so
now you have to input both target spheres of the elements to be
multiplied) and the logic was reworked.